tried to fix most things.
