+++ docker-compose.yaml
@@ -10,13 +10,10 @@ services:
       - client-build:/usr/src/client/dist
     ports:
       - "5173:5173"
-    secrets:
-      - db-password
     env_file:
       - .env
     environment:
       - VITE_SERVER_URL=http://server:3001
-      - NODE_ENV=development
     depends_on:
       server:
         condition: service_healthy
@@ -29,14 +26,14 @@ services:
       target: development
     volumes:
       - ./packages/server/src:/usr/src/server/src
+      - ./packages/server/migrations:/usr/src/server/migrations
       - client-build:/usr/src/client/dist
     ports:
       - "3001:3001"
     env_file:
       - .env
     environment:
-      - NODE_ENV=development
-      - POSTGRES_PASSWORD_FILE=/run/secrets/db-password
+      - DATABASE_URL=postgres://${POSTGRES_USER}:${POSTGRES_PASSWORD}@db:${POSTGRES_PORT}/${POSTGRES_DB}
     healthcheck:
       test: ["CMD", "curl", "-f", "http://server:3001/api/health"]
       interval: 10s
@@ -51,14 +48,10 @@ services:
     container_name: postgres
     restart: always
     user: postgres
-    secrets:
-      - db-password
     volumes:
       - db-data:/var/lib/postgresql/data
     env_file:
       - .env
-    environment:
-      - POSTGRES_PASSWORD_FILE=/run/secrets/db-password
     expose:
       - 5432
     healthcheck:
@@ -67,10 +60,24 @@ services:
       timeout: 5s
       retries: 5
 
+  pgadmin:
+    image: dpage/pgadmin4
+    restart: always
+    user: root
+    environment:
+      PGADMIN_DEFAULT_EMAIL: admin@admin.com
+      PGADMIN_DEFAULT_PASSWORD: root
+      PGADMIN_CONFIG_ENHANCED_COOKIE_PROTECTION: "False"
+      PGADMIN_CONFIG_CONSOLE_LOG_LEVEL: 40 # CRITICAL level
+      PGADMIN_CONFIG_FILE_LOG_LEVEL: 40 # Disable file logging
+    ports:
+      - 5050:80
+    volumes:
+      - pgadmin:/var/lib/pgadmin
+    logging:
+      driver: "none"
+
 volumes:
   db-data:
   client-build:
-
-secrets:
-  db-password:
-    file: db-secrets/password.txt
+  pgadmin:
