+++ package.json
@@ -16,26 +16,9 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@types/node": "^22.8.6",
-    "@types/react": "^18.3.12",
-    "@types/react-dom": "^18.3.1",
     "concurrently": "^9.0.1",
-    "file-loader": "^6.2.0",
     "npm-run-all": "^4.1.5",
     "prettier": "^3.3.3",
-    "prettier-plugin-tailwindcss": "^0.6.8",
-    "typescript": "^5.6.3"
-  },
-  "dependencies": {
-    "@react-email/components": "^0.0.25",
-    "axios": "^1.7.7",
-    "express-async-handler": "^1.2.0",
-    "next": "^14.2.15",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-email": "^3.0.1",
-    "resend": "^4.0.1-alpha.0",
-    "sharp": "^0.33.5",
-    "vite": "^5.4.9"
+    "prettier-plugin-tailwindcss": "^0.6.8"
   }
 }
