+++ packages/client/package.json
@@ -13,6 +13,8 @@
   "dependencies": {
     "@headlessui/react": "^2.1.10",
     "@heroicons/react": "^2.1.5",
+    "@react-email/components": "^0.0.26",
+    "axios": "^1.7.7",
     "bootstrap": "^5.3.3",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
