+++ packages/client/src/Components/Footer.tsx
@@ -1,8 +1,6 @@
-import React from "react";
-
-const Footer: React.FC = () => {
+export default function Footer() {
   return (
-    <footer className="relative border-lightGray top-[100vh] left-0 w-full border-t bg-white">
+    <footer className="border-lightGray relative left-0 top-[100vh] w-full border-t bg-white">
       <div className="mx-auto max-w-screen-xl px-6 py-2">
         <div className="flex flex-wrap items-start justify-between">
           <div className="mb-4 flex space-x-4">
@@ -86,5 +84,4 @@ const Footer: React.FC = () => {
       </div>
     </footer>
   );
-};
-export default Footer;
+}
