+++ packages/client/src/Components/MainContent.tsx
@@ -1,7 +1,6 @@
-import React from "react";
 import { Link } from "react-router-dom";
 
-const MainContent: React.FC = () => {
+export default function MainContent() {
   return (
     <div>
       <div className="relative h-[400px] w-full">
@@ -62,7 +61,7 @@ const MainContent: React.FC = () => {
 
       <div className="absolute inset-x-10 mt-32 flex h-[90%] flex-col justify-center">
         <p className="flex justify-center text-4xl">Why Choose Swift</p>
-        <div className="mt-14 ml-20 flex justify-center">
+        <div className="ml-20 mt-14 flex justify-center">
           <div className="relative h-[400px] w-full">
             <img
               src="../assets/images/boxes.jpg"
@@ -70,43 +69,43 @@ const MainContent: React.FC = () => {
               className="absolute h-auto max-h-[350px] w-auto max-w-[70%] rounded-md object-contain shadow-xl"
             />
           </div>
-          <div className="mt-4 ml-[40%] absolute flex flex-col items-center">
+          <div className="absolute ml-[40%] mt-4 flex flex-col items-center">
             <div className="flex space-x-48">
               <div className="relative flex flex-col items-center">
-                <div className="relative bg-mainGreen flex h-20 w-20 items-center justify-center rounded-full shadow-lg" />
+                <div className="bg-mainGreen relative flex h-20 w-20 items-center justify-center rounded-full shadow-lg" />
                 <img
                   src="../assets/images/fast.png"
                   alt="time"
-                  className="absolute top-[35px] left-[20%] h-[50%] w-[100%] object-fit"
+                  className="object-fit absolute left-[20%] top-[35px] h-[50%] w-[100%]"
                 />
                 <p className="mt-8 text-center">Fast Delivery</p>
               </div>
               <div className="relative flex flex-col items-center">
-                <div className="relative bg-mainGreen flex h-20 w-20 items-center justify-center rounded-full shadow-lg" />
+                <div className="bg-mainGreen relative flex h-20 w-20 items-center justify-center rounded-full shadow-lg" />
                 <img
                   src="../assets/images/tracking.png"
                   alt="tracking"
-                  className="absolute top-[35px] left-[55%] h-[50%] w-[48%] object-fit"
+                  className="object-fit absolute left-[55%] top-[35px] h-[50%] w-[48%]"
                 />
                 <p className="mt-8 text-center">Real Time Tracking</p>
               </div>
             </div>
             <div className="mt-16 flex space-x-40">
               <div className="relative flex flex-col items-center">
-                <div className="relative bg-mainGreen flex h-20 w-20 items-center justify-center rounded-full shadow-lg" />
+                <div className="bg-mainGreen relative flex h-20 w-20 items-center justify-center rounded-full shadow-lg" />
                 <img
                   src="../assets/images/earth.png"
                   alt="earth"
-                  className="absolute top-[35px] left-[45%] h-[50%] w-[48%] object-fit"
+                  className="object-fit absolute left-[45%] top-[35px] h-[50%] w-[48%]"
                 />
                 <p className="mt-8 text-center">Worldwide Shipping</p>
               </div>
               <div className="relative flex flex-col items-center">
-                <div className="relative bg-mainGreen flex h-20 w-20 items-center justify-center rounded-full shadow-lg" />
+                <div className="bg-mainGreen relative flex h-20 w-20 items-center justify-center rounded-full shadow-lg" />
                 <img
                   src="../assets/images/service.png"
                   alt="headset"
-                  className="absolute top-[35px] left-[45%] h-[50%] w-[40%] object-fit"
+                  className="object-fit absolute left-[45%] top-[35px] h-[50%] w-[40%]"
                 />
                 <p className="mt-8 text-center">24/7 Customer Support</p>
               </div>
@@ -116,6 +115,4 @@ const MainContent: React.FC = () => {
       </div>
     </div>
   );
-};
-
-export default MainContent;
+}
