+++ packages/client/src/Components/Navbar.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import { useState } from "react";
 import { Link } from "react-router-dom";
 
 interface NavItem {
@@ -12,7 +12,7 @@ interface NavbarProps {
   logo: string;
 }
 
-const Navbar: React.FC<NavbarProps> = ({ navItems, logo }) => {
+export default function Navbar({ navItems, logo }: NavbarProps) {
   const [openDropdown, setOpenDropdown] = useState<string | null>(null);
   const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
 
@@ -25,41 +25,36 @@ const Navbar: React.FC<NavbarProps> = ({ navItems, logo }) => {
   };
 
   return (
-    <nav className="flex flex-wrap items-center justify-between bg-white p-4 fixed top-0 z-10 w-full">
+    <nav className="fixed top-0 z-10 flex w-full flex-wrap items-center justify-between bg-white p-4">
       <div className="flex items-center">
-        <img src={logo} alt="Logo" className="h-10 w-auto mr-4" />
+        <img src={logo} alt="Logo" className="mr-4 h-10 w-auto" />
         <button
-          className="block md:hidden text-black focus:outline-none"
-          onClick={toggleMobileMenu}
-        >
-          {isMobileMenuOpen ? '✖' : '☰'}
+          className="block text-black focus:outline-none md:hidden"
+          onClick={toggleMobileMenu}>
+          {isMobileMenuOpen ? "✖" : "☰"}
         </button>
       </div>
 
       <div
-        className={`flex-col md:flex-row md:flex md:space-x-4 ${isMobileMenuOpen ? 'flex' : 'hidden'} md:flex`}
-      >
+        className={`flex-col md:flex md:flex-row md:space-x-4 ${isMobileMenuOpen ? "flex" : "hidden"} md:flex`}>
         {navItems.map((item) => (
           <div
             className="relative"
             key={item.name}
             onMouseEnter={() => toggleDropdown(item.name)}
-            onMouseLeave={() => toggleDropdown(null)}
-          >
+            onMouseLeave={() => toggleDropdown(null)}>
             <Link
               to={item.path}
-              className="text-black hover:bg-mainGreen px-4 py-2 rounded"
-            >
+              className="hover:bg-mainGreen rounded px-4 py-2 text-black">
               {item.name}
             </Link>
             {openDropdown === item.name && item.dropdownItems && (
-              <ul className="absolute left-0 bg-white text-black mt-2 shadow-lg z-10">
+              <ul className="absolute left-0 z-10 mt-2 bg-white text-black shadow-lg">
                 {item.dropdownItems.map((dropdownItem) => (
                   <li key={dropdownItem.name}>
                     <Link
                       to={dropdownItem.path}
-                      className="block px-4 py-2 hover:bg-mainGreen"
-                    >
+                      className="hover:bg-mainGreen block px-4 py-2">
                       {dropdownItem.name}
                     </Link>
                   </li>
@@ -71,6 +66,4 @@ const Navbar: React.FC<NavbarProps> = ({ navItems, logo }) => {
       </div>
     </nav>
   );
-};
-
-export default Navbar;
+}
