+++ packages/client/tsconfig.json
@@ -3,8 +3,5 @@
   "references": [
     { "path": "./tsconfig.app.json" },
     { "path": "./tsconfig.node.json" }
-  ],
-  "compilerOptions": {
-    "jsx": "react-jsx" // or "react" for older versions of React
-  }
+  ]
 }
